+++ sdk/eventhubs/azure-messaging-eventhubs-checkpointstore-blob/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Release History
+## 1.0.0-beta.4 (2019-12-02)
+- Artifact name changed from `preview` to `beta`.
+- `BlobCheckpointStore` implementation updated to match changes in `CheckpointStore` interface.
+
 ## 1.0.0-preview.3 (2019-11-01)
 - Renamed `BlobPartitionManager` to `BlobEventProcessorStore`
 - Added fully qualified namespace to list ownership API.
@@ -16,7 +20,7 @@ our efforts can be found in the [Azure SDK Design Guidelines for Java](https://a
 
 - Reactive streams support using [Project Reactor](https://projectreactor.io/).
 - Receive messages from all partitions of an Azure Event Hub using `EventProcessor`.
-- Provide an instance of `BlobEventProcessorStore` to your Event Processor. `BlobEventProcessorStore` uses Azure Blob Storage to 
+- Provide an instance of `BlobCheckpointStore` to your Event Processor. `BlobCheckpointStore` uses Azure Blob Storage to 
 store checkpoints and balance partition load among all instances of Event Processors.
 - Store checkpoint and partition ownership details in [Azure Storage Blobs](https://azure.microsoft.com/en-us/services/storage/blobs/).